Custom behaviors & improvements
* moved all errors to a separate file, as they are somewhat independent of every other types in the main file.
* added few ways to customize the behavior, via `ParseWithConfig()` function. It offers two behaviors now:
  * get envvar from something other than the actual envvar. this simplifies some of the test, as it no longer modifies the global envvar during the test run.
  * ability to specify the initial prefix. this is technically doable via the first, as you can pass in a fucntion that does `(key) -> wrapped(prefix + key)`
    but this offers somewhat simpler way to specify this.
